[!!!][FEATURE] Support for transient nodes
Introduces a new transient node type "NodeTemplate".
Both Node and NodeTemplate implement the (simplified) interface NodeInterface.
Node additionally implement the interface PersistedNodeInterface that contains
stubs related to persistence and context.

Besides this adds TypeConverters for both, persisted nodes and node templates.

This is a breaking change if you referred to NodeInterface in your code. In this case
you should replace the reference with PersistentNodeInterface – unless you only
invoke methods defined in the NodeInterface.

Change-Id: Id1e7b624ece93106ec028eda432fd3e323b86bff
Resolves: #44917
Resolves: #44918