[TASK] Move MATCH_PATTERN_NAME (back) to NodeInterface
In change Id1e7b624ece93106ec028eda432fd3e323b86bff that pattern was
moved to the PersistentNodeInterface, but NodeInterface instances still
have a name, thus the pattern should stay in NodeInterface.

This changes moves it back and adds checks against the pattern in the
setName() methods of Node and NodeTemplate.

Change-Id: Ic2e21de032def00c9248222cd0daa819b665ba28
Related: #44917