[BUGFIX] Add a check for empty property mapping for target types
While an property mapping for a target type is being made and
the property is empty an exception is thrown. This adds a check that
if the property it empty then do not try to convert it.

The problem occurs when for example trying to delete or publish an
image or image with text that has no image.

Change-Id: If0f13628a64cc37669aae94d3165358cc3fa36e5
Fixes: #46266
Fixes: #39987