[FEATURE] Support basic pagination for Node->getChildNodes()
This adds two parameters, $limit and $offset, to the getChildNodes()
method. In this implementation added and removed nodes which
have not yet been persisted can lead to diverging results. This
behavior seems to be acceptable though (for now).

This change set also introduces a Fluid View Helper Widget
which allows for paginating nodes in a template.

Change-Id: I83be69a7115d7a7de76b3955738b7309b2d483ad