[!!!][TASK] Context Cleanup
Nodes are now wrappers around NodeData entities and carry the
Context object. The same NodeData entity can be wrapped by
multiple Node objects with different contexts. Nodes should be
created with the NodeFactory only and contexts can be created
with the ContextFactory. NodeData entities are created
internally and are not public API.
There is no global Context and the NodeRepository
has no currentContext anymore.

Also implements a NodeConverter that is decoupled from the
TYPO3.Neos package, so that TYPO3CR has no dependencies on
Neos anymore.

Change-Id: Ic59a32c83b7f9f1e03c0f69f560bbb1795712e8c
Reviewed-on: https://review.typo3.org/21249
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek