[!!!][FEATURE] Add Node->getClosestAncestor($nodeType)
Now one can easily get the page a content element is located on by
calling::

    $contentNode->getClosestAncestor('TYPO3.Neos.NodeTypes:Page')

Using just getParent() does not work in most cases to get the page
the content element belongs to, because it is often inside a section
or other group of nodes.

This change is considered breaking as the method "findOnPath"
will now also return something if there are removed nodes
in the path. As this is currently not considered anywhere else
in the NodeRepository it shouldn't make a big difference.

Change-Id: Ia2f734c9657a3f0f04d5f1edca09206a92728333
Reviewed-on: https://review.typo3.org/19422
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller