[FEATURE] Node References
This change set introduces the possibility to refer to one or more other
nodes through a Node property. The new data types "reference" and
"references" can now be used for that purpose when declaring properties
of a node type.

'TYPO3.TYPO3CR:TestingNodeTypeWithReferences':
  properties:
    propertyPointingToOneNode:
      type: reference
    propertyPointingToMultipleNodes:
      type: references

Node::setProperty(), Node::getProperty() and Node::getProperties() all
support specifying node(s) through their Node Identifier or by passing
the actual Node object if the property type is declared as reference(s).
In case of "references" the value to pass is simply a numeric array of
Node objects or Node Identifiers.

Note that adjustments to Neos are necessary (separate commit) in order
to provide support for references in the user interface.

Change-Id: Iad700212e1ccb2da9a8bd162478e8cd9cb2fcb12
Reviewed-on: https://review.typo3.org/24167
Reviewed-by: Sebastian Kurfuerst
Tested-by: Sebastian Kurfuerst