[!!!][TASK] Refer to workspace instances rather than strings
The workspace model contains a lot of methods expecting
a workspace name (string) instead of Workspace instances.

This changes this by replacing $workspaceName with $workspace
where applicable.

This is a breaking change if you call one of the
``Workspace::publish*()`` methods in your code or if you created
slots for the recently added signals
``beforeNodePublishing`` and ``afterNodePublishing``

Change-Id: Iea48ec249f8ba74c9628478b7e671c093f45fedc
Reviewed-on: https://review.typo3.org/25141
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek