[FEATURE] Introduce flag for retrieving reference nodes as identifiers
This change introduces a flag for Node::getProperty() and
Node::getProperties() which changes their behaviour to return nodes
referenced in the properties as plain identifiers instead of actual
nodes.

This feature is necessary in order to export node references even if
these nodes (currently) don’t exist in the content repository.

Change-Id: I0f672088e3265b3d401009af86ef3c48dde45856
Reviewed-on: https://review.typo3.org/25238
Tested-by: Robert Lemke
Tested-by: Mr. Hudson
Reviewed-by: Aske Ertmann
Reviewed-by: Robert Lemke