[BUGFIX] Neos: Fatal error if referenced nodes are not visible
This fixes an issue caused by the Node->getProperty() method which
returned an array of referenced nodes (for properties of type "reference")
containing not only nodes but also "NULL" items. The root cause was that
the method could not handle referenced nodes which were not visible
according to the current content context.

Original error message:

PHP Fatal error:  Call to a member function getIdentifier() on a
non-object in TYPO3_Neos_Service_ContentElementWrappingService.php on
line 135"

Change-Id: Ibd7d8a21eb427e88518ed26b18d17435e7e3a250
Reviewed-on: https://review.typo3.org/25383
Tested-by: Robert Lemke
Tested-by: Mr. Hudson
Reviewed-by: Aske Ertmann
Reviewed-by: Bastian Waidelich
Reviewed-by: Robert Lemke