[BUGFIX] Make reference(s) properties self-repairing
It could happen that properties if type reference(s) end up in the
properties of NodeData not as identifiers (as expected) but as
serialized NodeData instances. This change repairs those cases
on-thy-fly.

Also it amends the check when settings reference(s) properties to also
handle AbstractNodeData in addition to NodeInterface.

Change-Id: I4ce32abf5a61075290075f197cddf899d975ddb5
Releases: master, 1.0
Reviewed-on: https://review.typo3.org/26259
Reviewed-by: Robert Lemke
Tested-by: Robert Lemke