[BUGFIX] Moved nodes must not be published directly
Changing the index of a node should not shine through to live
workspace until published.

The bug can be reproduced as follows:

* move a page in the navigate component behind a different page
  *on the same level*

Expected: The page should only be moved in the user workspace.
Actual: The page is moved directly in live workspace

This change fixes this behavior by materializing the node for every
move*() operation.

See I60313cc764f13c15847b1443dacaa6d365ad67cd for a related change.

Change-Id: I627e31a4701ea56dd123cc2090b685d35e437ba1
Reviewed-on: https://review.typo3.org/25711
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich