[!!!][TASK] Prevent multiple persistence commits on setPath
Currently nodes are persisted after every call of setPath, even
recursively which is unneeded. With this change persist happens
only at the end of a recursive setPath call chain.

This is only breaking if you use Node::setPath() in your own
code as persisting the changes is no longer happening
automatically.

Change-Id: Ie21be14ffaeb161232fcb631dc467c2040e2040a
Resolves: #53607
Reviewed-on: https://review.typo3.org/25749
Reviewed-by: Bastian Waidelich
Reviewed-by: Aske Ertmann
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller
Reviewed-by: Dominique Feyer
Tested-by: Dominique Feyer
Reviewed-by: Karsten Dambekalns
Tested-by: Karsten Dambekalns