[TASK] Add first level cache for nodes
A new FirstLevelNodeCache is implemented and used in the Context
and Node to serve identical queries for nodes from memory.

Needs Ic48370e9dcc752efa313a79fc8daf0cbc6761eaa in TYPO3.Neos for the
functional tests and Behat features to pass because of the changes in
Context.

Change-Id: Ic49af508acd55a8e6009afe0cf16a5acacfc963f
Releases: master
Reviewed-on: https://review.typo3.org/27504
Reviewed-by: Robert Lemke
Tested-by: Robert Lemke