[TASK] Remove ContextInterface
With the introduction of the 1st level cache for nodes the interface
would have needed the getFirstLevelCache() method - otherwise it would
have been useless.

OTOH this is rather internal and the existing context implementation is
pretty complex - I doubt implementing from scratch would be helpful.

This the ContextInterface is removed.

Change-Id: I0695f79fc7e0a70abc4660dc1354b85135885db4
Releases: master
Reviewed-on: https://review.typo3.org/27868
Reviewed-by: Robert Lemke
Tested-by: Robert Lemke